@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 
 
 
@@ -9,7 +9,7 @@ import { FormControl } from '@angular/forms';
   styleUrls: ['./input.component.css']
 })
 export class InputComponent implements OnInit {
-  @Input() control : FormControl = new FormControl();
+  @Input() control : UntypedFormControl = new UntypedFormControl();
   @Input() type = 'text';
   @Input() placeholder = "";
   @Input() format = '';
