@@ -6,11 +6,12 @@ import { TabComponent } from './tab/tab.component';
 import { InputComponent } from './input/input.component';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgxMaskModule } from 'ngx-mask';
+import { AlertComponent } from './alert/alert.component';
 
 // import { ModalService } from '../services/modal.service';
 
 @NgModule({
-  declarations: [ModalComponent, TabsContainerComponent, TabComponent, InputComponent],
+  declarations: [ModalComponent, TabsContainerComponent, TabComponent, InputComponent, AlertComponent],
   imports: [CommonModule, ReactiveFormsModule, NgxMaskModule.forRoot()],
   exports: [ModalComponent, TabComponent, TabsContainerComponent, InputComponent],
   // providers: [ModalService],
